Support pushing predicates of type varchar
This commit adds support for pushing predicates for varchar types. For
predicates on varchar types the node is of type `RelabelType` which
wraps the Var that we need for the predicate.

This commit ports commit 2c33c7a12c343e2e31ccce1f051b1be954d1831a from
external table to the foreign data wrapper code.