FDW: Add disable_ppd option
In external tables, there exists a mechanism to disable predicate
pushdown by setting a the `gp_external_enable_filter_pushdown` guc. The
PXF extension observes the `gp_external_enable_filter_pushdown` guc to
make sure predicate is not pushed down when filter pushdown is disabled.
For the PXF FDW extension, we provide the same functionality by adding a
new option for the server and foreign table definitions. We call this
option `disable_ppd`, and when the option is set to true, it will
disable predicate pushdown at the server or table level.